pyproject.toml
[tool.poetry]
name = "llm-sandbox"
version = "0.1.0"
description = "a sandbox for testing locally hosted LLM models"
authors = ["Brady Spiva <bspiva@redhat.com>"]
license = "Apache 2"
readme = "README.md"
packages = [{include = "llm_sandbox"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.186"
python-dotenv = "^1.0.0"
openai = "^0.27.7"
huggingface-hub = "^0.14.1"
tiktoken = "^0.4.0"
ipykernel = "^6.23.1"
pylint = "^2.17.4"
youtube-transcript-api = "^0.6.0"
pytube = "^15.0.0"
faiss-cpu = "^1.7.4"
wikipedia = "^1.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
